// pronti-frontend-main/front/src/app/modulos/impuesto-renta/form/form-impuesto-renta.enum.ts
export enum FormImpuestoRentaEnum {
  nombre = 'nombre',
  anio = 'anio',
  fraccionBasica1 = 'fraccionBasica1',
  impuestoFraccionBasica1 = 'impuestoFraccionBasica1',
  impuestoFraccionExcedente1 = 'impuestoFraccionExcedente1',
  fraccionBasica2 = 'fraccionBasica2',
  impuestoFraccionBasica2 = 'impuestoFraccionBasica2',
  impuestoFraccionExcedente2 = 'impuestoFraccionExcedente2',
  fraccionBasica3 = 'fraccionBasica3',
  impuestoFraccionBasica3 = 'impuestoFraccionBasica3',
  impuestoFraccionExcedente3 = 'impuestoFraccionExcedente3',
  fraccionBasica4 = 'fraccionBasica4',
  impuestoFraccionBasica4 = 'impuestoFraccionBasica4',
  impuestoFraccionExcedente4 = 'impuestoFraccionExcedente4',
  fraccionBasica5 = 'fraccionBasica5',
  impuestoFraccionBasica5 = 'impuestoFraccionBasica5',
  impuestoFraccionExcedente5 = 'impuestoFraccionExcedente5',
  fraccionBasica6 = 'fraccionBasica6',
  impuestoFraccionBasica6 = 'impuestoFraccionBasica6',
  impuestoFraccionExcedente6 = 'impuestoFraccionExcedente6',
  fraccionBasica7 = 'fraccionBasica7',
  impuestoFraccionBasica7 = 'impuestoFraccionBasica7',
  impuestoFraccionExcedente7 = 'impuestoFraccionExcedente7',
  fraccionBasica8 = 'fraccionBasica8',
  impuestoFraccionBasica8 = 'impuestoFraccionBasica8',
  impuestoFraccionExcedente8 = 'impuestoFraccionExcedente8',
  fraccionBasica9 = 'fraccionBasica9',
  impuestoFraccionBasica9 = 'impuestoFraccionBasica9',
  impuestoFraccionExcedente9 = 'impuestoFraccionExcedente9',
  fraccionBasica10 = 'fraccionBasica10',
  impuestoFraccionBasica10 = 'impuestoFraccionBasica10',
  impuestoFraccionExcedente10 = 'impuestoFraccionExcedente10',

}